Move components to @modules and @primitives
